+++ Backend/WorldData/EconomyManager.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using JuniorProject.Backend.Agents;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,46 +7,39 @@ using System.Threading.Tasks;
 
 namespace JuniorProject.Backend.WorldData
 {
-    internal class EconomyManager
+    class EconomyManager
     {
-        int currentGold;
-        int currentWood;
-        int currentIron;
-        int currentStone;
-        int currentFood;
+        int woodValue;
+        int ironValue;
+        int stoneValue;
+        int foodValue;
 
-        int woodDemand;
-        int ironDemand;
-        int stoneDemand;
-        int foodDemand;
+        public EconomyManager()
+        {
+            woodValue = 100;
+            ironValue = 100;
+            stoneValue = 100;
+            foodValue = 100;
+        }
 
-        void SetWoodDemand() 
+        void CalcWoodPrice(int woodDemand)
         {
-            if (currentWood <= 100 && woodDemand < 10)
-                woodDemand++;
-            else if (currentWood >= 500 && woodDemand > 0)
-                woodDemand--;
+            woodValue *= ((woodDemand - 5) / 20) + 1;
         }
-        void SetIronDemand()
+
+        void CalcIronPrice(int ironDemand)
         {
-            if (currentIron <= 100 && ironDemand < 10)
-                ironDemand++;
-            else if (currentIron >= 500 && ironDemand > 0)
-                ironDemand--;
+            ironValue *= ((ironDemand - 5) / 20) + 1;
         }
-        void SetStoneDemand()
+
+        void CalcStonePrice(int stoneDemand)
         {
-            if (currentStone <= 100 && stoneDemand < 10)
-                stoneDemand++;
-            else if (currentStone >= 500 && stoneDemand > 0)
-                stoneDemand--;
+            stoneValue *= ((stoneDemand - 5) / 20) + 1;
         }
-        void SetFoodDemand()
+
+        void CalcFoodPrice(NationResources buyer)
         {
-            if (currentFood <= 100 && foodDemand < 10)
-                foodDemand++;
-            else if (currentFood >= 500 && foodDemand > 0)
-                foodDemand--;
+            foodValue *= ((buyer.foodDemand - 5) / 20) + 1;
         }
     }
 }
